Flytt henting av totrinn til egen context for å kunne bruke statusen flere steder
