+++ src/frontend/Sider/Behandling/Totrinnskontroll/Totrinnskontroll.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { FC, useEffect, useState } from 'react';
+import { FC, useState } from 'react';
 
 import { useNavigate } from 'react-router-dom';
 import styled from 'styled-components';
@@ -10,12 +10,8 @@ import FatteVedtak from './FatteVedtak';
 import SendtTilBeslutter from './SendtTilBeslutter';
 import TotrinnskontrollUnderkjent from './TotrinnskontrollUnderkjent';
 import { TotrinnskontrollResponse, TotrinnskontrollStatus } from './typer';
-import { useBehandling } from '../../../context/BehandlingContext';
-import { useHentTotrinnskontroll } from '../../../hooks/useHentTotrinnskontroll';
-import { usePrevious } from '../../../hooks/usePrevious';
+import { useTotrinnskontroll } from '../../../context/TotrinnskontrollContext';
 import { ModalWrapper } from '../../../komponenter/Modal/ModalWrapper';
-import { Behandling } from '../../../typer/behandling/behandling';
-import { BehandlingStatus } from '../../../typer/behandling/behandlingStatus';
 import { Ressurs, RessursStatus } from '../../../typer/ressurs';
 
 const BorderBox = styled.div`
@@ -56,41 +52,12 @@ const TotrinnskontrollSwitch: FC<{
     }
 };
 
-/**
- * Skal hente totrinnskontroll:
- *  * Første rendering av siden hvis status er UTREDES eller FATTER_VEDTAK
- *  * Hvis status går fra utredes til fatter vedtak (sender til totrinnskontroll)
- */
-const skalHenteTotrinnskontroll = (
-    prevBehandling: Behandling | undefined,
-    behandling: Behandling
-) => {
-    const forrigeStatus = prevBehandling?.status;
-    const nyStatus = behandling.status;
-    const statusErUendret = forrigeStatus === nyStatus;
-
-    if (statusErUendret) return false;
-    // Init-henting / sender til beslutter (send til beslutter kan erstattes hvis den knappen blir flyttet fra brev)
-    if (nyStatus === BehandlingStatus.FATTER_VEDTAK) return true;
-
-    // Init-henting
-    return !forrigeStatus && nyStatus === BehandlingStatus.UTREDES;
-};
 const Totrinnskontroll: FC = () => {
     const navigate = useNavigate();
-    const { behandling } = useBehandling();
-    const prevBehandling = usePrevious(behandling);
 
     const [visGodkjentModal, settVisGodkjentModal] = useState(false);
 
-    const { totrinnskontroll, hentTotrinnskontroll, settTotrinnskontroll } =
-        useHentTotrinnskontroll();
-
-    useEffect(() => {
-        if (skalHenteTotrinnskontroll(prevBehandling, behandling)) {
-            hentTotrinnskontroll(behandling.id);
-        }
-    }, [prevBehandling, behandling, hentTotrinnskontroll]);
+    const { totrinnskontroll, settTotrinnskontroll } = useTotrinnskontroll();
 
     const skalViseTotrinnskontrollSwitch =
         totrinnskontroll.status === RessursStatus.SUKSESS &&
